Better-motivated renaming of packages
This patch aims to make the KDL packages more similar in naming
convention to the other packages in the Orocos toolchain. It also
resolves an issue with catkin naming conventions which require that
packages be under_scored.

Additionally, the metapackage is now given the long name corresponding
to the repository, and the actual KDL package is just called `kdl`
furthermore, the python KDL bindings, which are exposed as "PyKDL" in
Python, has been renamed to "pykdl".