KDL Frames: Added Improper Rotation test for GetRotAngle
KDL assumes that rotation matrices are proper, i.e. orthonormal with an eigenvalue of +1.  Improper rotation matrices with an eigenvalue of -1, such as the negative identity matrix, involve a mirroring of axes such as going from a right-hand to left-hand coordinate system or vice versa.  However, rotation matrices are always proper if operating within a right hand or left hand system.  This means also that GetRotAngle will not return the correct angle-axis if the rotation matrix is improper.  Two test cases were added to verify this, one for a nonsingular case and the other for a singular case (180 deg rotation) to demonstrate this.