KDL frames:  removed unnecessary epsilon checks in GetRotAngle singular case
The set of conditionals where xx, yy, and zz are checked for being < epsilon seems to be searching for the case where the matrix is the negative identity matrix.  For example, if xx > yy and xx > zz and xx < epsilon, then the diagonal elements data[0], data[4], and data[8] are all approximately equal -1.  Yet this will return an angle-axis of (0, 1/sqrt2, 1/sqrt2) which corresponds to a rotation matrix of [-1 0 0; 0 0 1; 0 1 0] which is not the negative identity matrix.  The same is true for yy or zz being the largest value but still ~0.  But the negative identity matrix is an improper rotation matrix so the checks should not be there.

This code was drawn from the java code at http://www.euclideanspace.com/maths/geometry/rotations/conversions/matrixToAngle/index.htm. In the final entry under “Correspondence on this page”, Gene Gorokhovsky basically gives protocode for the 180 deg singularity case, which does not include the ‘ < epsilon ‘ checks.  So it seems this was added afterwards as a way to prevent a divide by zero for the case of the negative identity matrix which is not a legal input.