Implement LDL solver for finding joint acceleration
The joint acceleration is found by multiplying all of the joint torques by the inverse of the joint space mass matrix.  Rather than explicitly calculating the inverse of the mass matrix, this can be calculated more efficiently by doing a LDL^T decomposition of the mass matrix and multiplying by the torques in one operation.  This is similar to a Cholesky Decomposition but avoids the need for calculating a square root.  Size checks were added for all of the matrices and vectors based on the row size of input square matrix A.  Note that the upper triangular elements of L are set to 0 and the diagonal elements are set to 1 so that the output value of L is correct.  The algorithm does not do this automatically because it is not required to compute the output vector.