Add FD solver tests
Added a “development” test for the FD solver that outlines how the new class is structured and does checks at each step.  Also added a “consistency” test for the FD solver that takes the acceleration output of the FD solver and inputs it to the ID solver and checks whether the output torque agrees with the input of the FD solver.