Remove python_orocos_kdl env-hooks, support devel-space build of python_orocos_kdl and invoke catkin_package()
Macro catkin_package() takes care of installing package.xml and setup scripts.
Usually it also installs cmake and pkg-config configuration files, but we skip this because
orocos_kdl already provides its own configuration and the Python bindings do not require it.

If the Python module is built in the correct path inside the devel-space, pointed to
by PYTHON_INSTALL_DIR and CATKIN_GLOBAL_PYTHON_DESTINATION, and also installed to this
directory, no env-hooks are required to update the PYTHONPATH and python_orocos_kdl even works
in devel-space.

This patch also allows to build orocos_kinematics_dynamics with colcon and colcon-ros (for ROS 1),
because python_orocos_kdl is correctly detected as a catkin package and colcon will add the
necessary hooks to update the PYTHONPATH, too.