nr pos solver: added degraded check and get/set functions for parameters
The position solver was causing an IK failure if it did not converge after the maximum number of iterations was exceeded.  However, when singularity avoidance is active (degraded), the solution may not converge but the IK should be allowed to continue.  So a degraded flag was added and is checked when the maximum number of iterations is exceeded.  If it is degraded, the error is set to E_DEGRADED to allow it to continue.  Otherwise, it is set to E_MAX_ITERATIONS_EXCEEDED which will cause the IK to fail.  Added functions to set the input parameters (maxiter, eps) and to retrieve the twist error, number of iterations performed, and the velocity solver status.