pinv vel solver:  added sort function for singular values
The singular value decomposition in SVD_HH does not sort the singular values in descending order.  So a sort function was added to do this based on the same code used in SVD_eigen_HH.  The sorted version of S called Smaxtomin is then used to find the minimum of the six largest singular values and is also what is returned by the get Sigma function used to calculate the manipulability.