Fix CMake warning on Windows
Because the minimum cmake version is set to 3.0.2, we see the following policy warning on Windows:

   CMake Warning (dev) at C:/ci/ws/install/share/cmake/pybind11/pybind11Common.cmake:127 (if):
     Policy CMP0054 is not set: Only interpret if() arguments as variables or
     keywords when unquoted.  Run "cmake --help-policy CMP0054" for policy
     details.  Use the cmake_policy command to set the policy and suppress this
     warning.

     Quoted variables like "MSVC" will no longer be dereferenced when the policy
     is set to NEW.  Since the policy is not set the OLD behavior will be used.

If we want to continue to support older versions of CMake, we should opt into the new behavior to
not break the logic in pybind11Common.cmake. Alternatively, we could increase the miminum cmake
version of this project.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>