Rename variables with reserved names
From the GNU libc manual[0]:

In addition to the names documented in this manual, reserved names
include all external identifiers (global functions and variables)
that begin with an underscore (‘_’) and all identifiers regardless
of use that begin with either two underscores or an underscore
followed by a capital letter are reserved names. This is so that
the library and header files can define functions, variables, and
macros for internal purposes without risk of conflict with names
in user programs.

[0] https://www.gnu.org/software/libc/manual/html_node/Reserved-Names.html