cmake: Allow PYTHON_SITE_PACKAGES_INSTALL_DIR set extern (#461)
The variable PYTHON_SITE_PACKAGES_INSTALL_DIR is kind of
hardcoded in the CMakeLists.txt to .../dist-packages. This
is not true for all linux distros. dist-packages comes
from debian and debian like distros. Python core uses
site-packages.

We use an unintrusive way and allow the variable to be
passed from cmake, so you at least have the option to
set a different install location (i.e. site-packages)
from extern.

Signed-off-by: Matthias Schoepfer <matthias.schoepfer@googlemail.com>